+++ packages/react-i18n/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Fixed types on `translate` when `undefined` or `boolean` is used in a replacement dictionary.
 
 ## 5.3.4 - 2021-04-13